search_service/requirements.txt
django
djangorestframework
python-decouple
drf-spectacular
kafka-python
django-elasticsearch-dsl
django-elasticsearch-dsl-drf